+++ tfe/data_source_policy_set.go
@@ -36,7 +36,6 @@ func dataSourceTFEPolicySet() *schema.Resource {
 			"kind": {
 				Type:     schema.TypeString,
 				Optional: true,
-				Default:  string(tfe.Sentinel),
 			},
 
 			"overridable": {